bi_api/requirements.txt
fastapi==0.119.1
uvicorn==0.37.0
pydantic==2.12.2
python-dotenv==1.1.1
openai==1.109.1
openai-agents==0.3.3
mcp==1.17.0
supabase==2.22.0
requests==2.32.5
python-multipart==0.0.20
